zyra-ml/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.0
python-dotenv==1.0.0
pydantic==2.10.0
joblib==1.4.0
shap==0.46.0
matplotlib==3.9.0
